+++ libs/common/src/scanner/scanner.ts
